refactor: move reducer to seperate dir
